upgrade vec3 functions

+++ src/vec3/vec3_length_squared.c
@@ -12,9 +12,9 @@
 
 #include "struct.h"
 
-double	vec3_length_squared(const t_vec3 *v);
+float	vec3_length_squared(const t_vec3 *v);
 
-inline double	vec3_length_squared(const t_vec3 *v)
+inline float	vec3_length_squared(const t_vec3 *v)
 {
 	return (v->x * v->x + v->y * v->y + v->z * v->z);
 }
